# openapi.yaml
openapi: 3.1.0
info:
  title: Acoboo API - OpenAPI 3.0
  description: |-
    家計簿アプリケーションAcobooのAPIドキュメントです。
    Powered by takashi@misskey.systems
  version: 1.0.0
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: https://acoboo.swagger.io/api/v1
tags:
  - name: monetary events
    description: 収入 / 支出
  - name: accounts
    description: 口座
  - name: categories
    description: カテゴリー

paths:
  /monetary-events:
    post:
      tags:
        - monetary events
      summary: 新しい収入/支出を登録します
      description: 新しい収入/支出を登録します
      operationId: addMonetaryEvent
      requestBody:
        description: 新しい収入/支出を登録します
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonetaryEvent"
          application/xml:
            schema:
              $ref: "#/components/schemas/MonetaryEvent"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/MonetaryEvent"
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonetaryEvent"
            application/xml:
              schema:
                $ref: "#/components/schemas/MonetaryEvent"
        "400":
          description: Invalid input
        "422":
          description: Validation exception

    get:
      tags:
        - monetary events
      summary: 収入/支出の一覧を取得します
      description: 収入/支出の一覧を取得します
      operationId: getMonetaryEvents
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonetaryEvents"
            application/xml:
              schema:
                $ref: "#/components/schemas/MonetaryEvents"
        "400":
          description: Invalid input
        "422":
          description: Validation exception
  /monetary-events/{id}:
    get:
      tags:
        - monetary events
      summary: Find monetary event by ID
      description: Returns a single monetary event
      operationId: getMonetaryEventById
      parameters:
        - name: id
          in: path
          description: ID of monetary event to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonetaryEvent"
            application/xml:
              schema:
                $ref: "#/components/schemas/MonetaryEvent"
        "400":
          description: Invalid ID supplied
        "404":
          description: Monetary event not found
    post:
      tags:
        - monetary events
      summary: Updates a monetary events in the store with form data
      description: ""
      operationId: updateMonetaryEvent
      parameters:
        - name: id
          in: path
          description: ID of monetary event that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
        - name: name
          in: query
          description: Name of monetary event that needs to be updated
          schema:
            type: string
        - name: status
          in: query
          description: Status of monetary event that needs to be updated
          schema:
            type: string
      responses:
        "400":
          description: Invalid input
    delete:
      tags:
        - monetary events
      summary: Deletes a monetary event
      description: delete a monetary event
      operationId: deletemonetary event
      parameters:
        - name: id
          in: path
          description: monetary event id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "400":
          description: Invalid monetary event value

  /accounts:
    post:
      tags:
        - accounts
      summary: 新しい口座を登録します
      description: 新しい口座を登録します
      operationId: addAccount
      requestBody:
        description: 新しい口座を登録します
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Account"
          application/xml:
            schema:
              $ref: "#/components/schemas/Account"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Account"
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
            application/xml:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          description: Invalid input
        "422":
          description: Validation exception

    get:
      tags:
        - accounts
      summary: 口座の一覧を取得します
      description: 口座の一覧を取得します
      operationId: getAccounts
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
            application/xml:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          description: Invalid input
        "422":
          description: Validation exception
  /accounts/{id}:
    get:
      tags:
        - accounts
      summary: Find account by ID
      description: Returns a single account
      operationId: getAccountById
      parameters:
        - name: id
          in: path
          description: ID of account to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonetaryEvent"
            application/xml:
              schema:
                $ref: "#/components/schemas/MonetaryEvent"
        "400":
          description: Invalid ID supplied
        "404":
          description: account not found
    post:
      tags:
        - accounts
      summary: Updates a accounts in the store with form data
      description: ""
      operationId: updateAccount
      parameters:
        - name: id
          in: path
          description: ID of account that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
        - name: name
          in: query
          description: Name of account that needs to be updated
          schema:
            type: string
        - name: status
          in: query
          description: Status of account that needs to be updated
          schema:
            type: string
      responses:
        "400":
          description: Invalid input
    delete:
      tags:
        - accounts
      summary: Deletes a account
      description: delete a account
      operationId: deleteAccount
      parameters:
        - name: id
          in: path
          description: account id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "400":
          description: Invalid account value

  /categories:
    post:
      tags:
        - categories
      summary: 新しいカテゴリーを登録します
      description: 新しいカテゴリーを登録します
      operationId: addCategory
      requestBody:
        description: 新しいカテゴリーを登録します
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Category"
          application/xml:
            schema:
              $ref: "#/components/schemas/Category"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Category"
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
            application/xml:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Invalid input
        "422":
          description: Validation exception

    get:
      tags:
        - categories
      summary: カテゴリーの一覧を取得します
      description: カテゴリーの一覧を取得します
      operationId: getCategories
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Categories"
            application/xml:
              schema:
                $ref: "#/components/schemas/Categories"
        "400":
          description: Invalid input
        "422":
          description: Validation exception
  /categories/{id}:
    get:
      tags:
        - categories
      summary: Find cateogry by ID
      description: Returns a single cateogry
      operationId: getEventById
      parameters:
        - name: id
          in: path
          description: ID of category to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
            application/xml:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Invalid ID supplied
        "404":
          description: category not found
    post:
      tags:
        - categories
      summary: Updates a category in the store with form data
      description: ""
      operationId: updateCategory
      parameters:
        - name: id
          in: path
          description: ID of category that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
        - name: name
          in: query
          description: Name of category that needs to be updated
          schema:
            type: string
        - name: status
          in: query
          description: Status of category that needs to be updated
          schema:
            type: string
      responses:
        "400":
          description: Invalid input
    delete:
      tags:
        - categories
      summary: Deletes a category
      description: delete a category
      operationId: deleteEvent
      parameters:
        - name: id
          in: path
          description: category id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "400":
          description: Invalid category value

components:
  schemas:
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
          x-oapi-codegen-extra-tags:
            gorm: primaryKey
        name:
          type: string
          example: 食費
      xml:
        name: category
    Categories:
      type: array
      items:
        $ref: "#/components/schemas/Category"
    Account:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
          x-oapi-codegen-extra-tags:
            gorm: primaryKey
        name:
          type: string
          example: テスト銀行
      xml:
        name: account
    MonetaryEvent:
      required:
        - name
        - eventType
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
          x-oapi-codegen-extra-tags:
            gorm: primaryKey
        name:
          type: string
          example: スーパーで買い物
        eventType:
          type: string
          description: 収入/支出
          enum:
            - income
            - expense
        category:
          $ref: "#/components/schemas/Category"
        account:
          $ref: "#/components/schemas/Account"
    MonetaryEvents:
      type: array
      items:
        $ref: "#/components/schemas/MonetaryEvent"
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
      xml:
        name: "##default"

  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header